requirements.txt
numpy==2.1.2
scipy==1.14.1
pillow==11.0.0
matplotlib==3.9.2